requirements.txt
Flask==2.3.2
Flask_SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
scikit-learn==1.2.2
flask-cors==3.0.10
flask-restx==1.1.0
pytest==7.4.0
flake8==6.0.0
